action creactor를 사용하여 리팩터링
