+++ src/App.jsx
@@ -1,60 +1,50 @@
-import { useState } from 'react';
-
 import { useDispatch, useSelector } from 'react-redux';
 
 import Page from './Page';
 
-const initialState = {
-  newId: 100,
-  taskTitle: '',
-  tasks: [],
-};
-
-function updateTaskTitle(state, taskTitle) {
+// action creator
+function updateTaskTitle(taskTitle) {
   return {
-    ...state,
-    taskTitle,
+    type: 'updateTaskTitle',
+    payload: {
+      taskTitle,
+    },
   };
 }
 
-function addTask(state) {
-  const { newId, taskTitle, tasks } = state;
-
+function addTask() {
   return {
-    ...state,
-    newId: newId + 1,
-    taskTitle: '',
-    tasks: [...tasks, { id: newId, title: taskTitle }],
+    type: 'addTask',
   };
 }
 
-function deleteTask(state, id) {
-  const { tasks } = state;
-
+function deleteTask(id) {
   return {
-    ...state,
-    tasks: tasks.filter((task) => task.id !== id),
+    type: 'deleteTask',
+    payload: {
+      id,
+    },
   };
 }
 
 export default function App() {
-  const dispatch = useDispatch();
-
   const { taskTitle, tasks } = useSelector((state) => ({
     taskTitle: state.taskTitle,
     tasks: state.tasks,
   }));
 
+  const dispatch = useDispatch();
+
   function handleChangeTitle(event) {
-    dispatch(updateTaskTitle(state, event.target.value));
+    dispatch(updateTaskTitle(event.target.value));
   }
 
   function handleClickAddTask() {
-    dispatch(addTask(state));
+    dispatch(addTask());
   }
 
   function handleClickDeleteTask(id) {
-    dispatch(deleteTask(state, id));
+    dispatch(deleteTask(id));
   }
 
   return (
